Updated included boost version from 1.55 to 1.59, updated vs2008 project and cmake build files, both working. - Left to do: Xcode, Code.blocks project files, rebuild on mac+linux. - also fixed a typo in the header docs and removed some stray dependencies from SendStringMarkersC vs2008 project file

+++ LSL/liblsl/external/lslboost/thread/detail/thread_group.hpp
@@ -85,7 +85,7 @@ namespace lslboost
             if(thrd)
             {
                 BOOST_THREAD_ASSERT_PRECONDITION( ! is_thread_in(thrd) ,
-                    thread_resource_error(system::errc::resource_deadlock_would_occur, "lslboost::thread_group: trying to add a duplicated thread")
+                    thread_resource_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost::thread_group: trying to add a duplicated thread")
                 );
 
                 lslboost::lock_guard<shared_mutex> guard(m);
@@ -106,7 +106,7 @@ namespace lslboost
         void join_all()
         {
             BOOST_THREAD_ASSERT_PRECONDITION( ! is_this_thread_in() ,
-                thread_resource_error(system::errc::resource_deadlock_would_occur, "lslboost::thread_group: trying joining itself")
+                thread_resource_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost::thread_group: trying joining itself")
             );
             lslboost::shared_lock<shared_mutex> guard(m);
