+++ LSL/liblsl/external/lslboost/thread/lock_types.hpp
@@ -336,12 +336,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if (owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       m->lock();
       is_locked = true;
@@ -351,12 +351,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if (owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked = m->try_lock();
       return is_locked;
@@ -367,11 +367,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked=m->timed_lock(relative_time);
       return is_locked;
@@ -381,11 +381,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked=m->timed_lock(absolute_time);
       return is_locked;
@@ -394,11 +394,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked=m->timed_lock(absolute_time);
       return is_locked;
@@ -411,11 +411,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked=m->try_lock_for(rel_time);
       return is_locked;
@@ -425,11 +425,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost unique_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost unique_lock owns already the mutex"));
       }
       is_locked=m->try_lock_until(abs_time);
       return is_locked;
@@ -441,12 +441,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock has no mutex"));
       }
       if (!owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost unique_lock doesn't own the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost unique_lock doesn't own the mutex"));
       }
       m->unlock();
       is_locked = false;
@@ -636,11 +636,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       m->lock_shared();
       is_locked=true;
@@ -649,11 +649,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->try_lock_shared();
       return is_locked;
@@ -663,11 +663,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->timed_lock_shared(target_time);
       return is_locked;
@@ -677,11 +677,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->timed_lock_shared(target_time);
       return is_locked;
@@ -693,11 +693,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->try_lock_shared_for(rel_time);
       return is_locked;
@@ -707,11 +707,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->try_lock_shared_until(abs_time);
       return is_locked;
@@ -721,11 +721,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(!owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock doesn't own the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock doesn't own the mutex"));
       }
       m->unlock_shared();
       is_locked=false;
@@ -934,12 +934,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if (owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost upgrade_lock owns already the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost upgrade_lock owns already the mutex"));
       }
       m->lock_upgrade();
       is_locked = true;
@@ -949,12 +949,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if (owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost upgrade_lock owns already the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost upgrade_lock owns already the mutex"));
       }
       is_locked = m->try_lock_upgrade();
       return is_locked;
@@ -964,12 +964,12 @@ namespace lslboost
       if (m == 0)
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if (!owns_lock())
       {
         lslboost::throw_exception(
-            lslboost::lock_error(system::errc::operation_not_permitted, "lslboost upgrade_lock doesn't own the mutex"));
+            lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost upgrade_lock doesn't own the mutex"));
       }
       m->unlock_upgrade();
       is_locked = false;
@@ -980,11 +980,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->try_lock_upgrade_for(rel_time);
       return is_locked;
@@ -994,11 +994,11 @@ namespace lslboost
     {
       if(m==0)
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::operation_not_permitted, "lslboost shared_lock has no mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::operation_not_permitted), "lslboost shared_lock has no mutex"));
       }
       if(owns_lock())
       {
-        lslboost::throw_exception(lslboost::lock_error(system::errc::resource_deadlock_would_occur, "lslboost shared_lock owns already the mutex"));
+        lslboost::throw_exception(lslboost::lock_error(static_cast<int>(system::errc::resource_deadlock_would_occur), "lslboost shared_lock owns already the mutex"));
       }
       is_locked=m->try_lock_upgrade_until(abs_time);
       return is_locked;
